SMC-4550: Nearly finished unit tests

+++ test/unit/security/lookup-tables.js
@@ -12,6 +12,7 @@ describe(test.testName(__filename, 3), function() {
   const util = require('util');
   let mockErrorService;
   let dbPath = path.resolve(__dirname, '../../__fixtures/test/test_lookup_db');
+
   before('Sets up errorService', done => {
     const ErrorService = require('../../../lib/services/error/service');
     mockErrorService = new ErrorService({
@@ -36,8 +37,9 @@ describe(test.testName(__filename, 3), function() {
     mockDataService.__updateDatabaseVersions = async () => {};
     mockDataService.initialize({ filename: dbPath }, done);
   });
+
   after('deletes test DB', done => {
-    fs.unlinkSync(dbPath);
+    if (fs.existsSync(dbPath)) fs.unlinkSync(dbPath);
     done();
   });
 
@@ -52,6 +54,7 @@ describe(test.testName(__filename, 3), function() {
     test.expect(lookupTables).to.be.ok();
     done();
   });
+
   it('Can initialize an instance of LookupTables', done => {
     let lookupTables = LookupTables.create();
     lookupTables.initialize(mockHappn);
@@ -62,9 +65,9 @@ describe(test.testName(__filename, 3), function() {
   it('Can upsert a path into a lookupTable (creatng table)', async () => {
     let lookupTables = LookupTables.create();
     lookupTables.initialize(mockHappn);
-    await lookupTables.insertPath('/1/2/3', 'newlookuptable');
+    await lookupTables.insertPath('/1/2/3', 'newUpsertTable');
     let stored = await mockHappn.services.data.get(
-      '/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable/1/2/3'
+      '/_SYSTEM/_SECURITY/_LOOKUP/newUpsertTable/1/2/3'
     );
     test.expect(stored.data).to.eql({ authorized: true });
   });
@@ -73,14 +76,14 @@ describe(test.testName(__filename, 3), function() {
     let lookupTables = LookupTables.create();
     lookupTables.initialize(mockHappn);
     let table = {
-      name: 'newlookuptable2',
+      name: 'upsertFullTable',
       paths: ['1/2/3', '/4/5/6', '/7/8/9'] //Inconsitency deliberate
     };
     await lookupTables.upsertLookupTable(table);
     for (let lookupPath of table.paths) {
       let testPath = lookupPath.replace(/^\//, '');
       let stored = await mockHappn.services.data.get(
-        `/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable2/${testPath}`
+        `/_SYSTEM/_SECURITY/_LOOKUP/upsertFullTable/${testPath}`
       );
       test.expect(stored.data).to.eql({ authorized: true });
     }
@@ -90,22 +93,18 @@ describe(test.testName(__filename, 3), function() {
     let lookupTables = LookupTables.create();
     lookupTables.initialize(mockHappn);
     let table = {
-      name: 'newlookuptable3',
+      name: 'upsertNewPath',
       paths: ['1/2/3', '4/5/6', '7/8/9']
     };
     await lookupTables.upsertLookupTable(table);
     let stored = await mockHappn.services.data.get(
-      '/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable3/1/2/3'
+      '/_SYSTEM/_SECURITY/_LOOKUP/upsertNewPath/1/2/3'
     );
     test.expect(stored.data).to.eql({ authorized: true });
-    stored = await mockHappn.services.data.get(
-      '/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable3/10/11/12'
-    );
+    stored = await mockHappn.services.data.get('/_SYSTEM/_SECURITY/_LOOKUP/upsertNewPath/10/11/12');
     test.expect(stored).to.not.be.ok();
-    await lookupTables.insertPath('/10/11/12', 'newlookuptable3');
-    stored = await mockHappn.services.data.get(
-      '/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable3/10/11/12'
-    );
+    await lookupTables.insertPath('/10/11/12', 'upsertNewPath');
+    stored = await mockHappn.services.data.get('/_SYSTEM/_SECURITY/_LOOKUP/upsertNewPath/10/11/12');
     test.expect(stored.data).to.eql({ authorized: true });
   });
 
@@ -113,16 +112,14 @@ describe(test.testName(__filename, 3), function() {
     let lookupTables = LookupTables.create();
     lookupTables.initialize(mockHappn);
     let table = {
-      name: 'newlookuptable4',
+      name: 'removeTable',
       paths: ['1/2/3', '4/5/6', '7/8/9']
     };
     await lookupTables.upsertLookupTable(table);
-    let stored = await mockHappn.services.data.get(
-      '/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable4/1/2/3'
-    );
+    let stored = await mockHappn.services.data.get('/_SYSTEM/_SECURITY/_LOOKUP/removeTable/1/2/3');
     test.expect(stored.data).to.eql({ authorized: true });
 
-    await lookupTables.removePath('/1/2/3', 'newlookuptable4');
+    await lookupTables.removePath('/1/2/3', 'removeTable');
     await wait(500);
     stored = await mockHappn.services.data.get('/_SYSTEM/_SECURITY/_LOOKUP/newlookuptable4/1/2/3');
     test.expect(stored).to.not.be.ok();
@@ -138,7 +135,7 @@ describe(test.testName(__filename, 3), function() {
     let lookupTables = LookupTables.create();
     let tableName = 'tableName';
     let path = '/domain/component/tableName/1/2/3/4';
-    test.expect(lookupTables.extractPath(path, tableName)).to.eql('1/2/3/4');
+    test.expect(lookupTables.__extractPath(path, tableName)).to.eql('1/2/3/4');
     done();
   });
 
@@ -147,17 +144,21 @@ describe(test.testName(__filename, 3), function() {
     let lookupTables = LookupTables.create();
     let tableName = 'tableName';
     let path = '/domain/component/tableName/1/2/3/tableName/5/6/7';
-    test.expect(lookupTables.extractPath(path, tableName)).to.be('1/2/3/tableName/5/6/7');
+    test.expect(lookupTables.__extractPath(path, tableName)).to.be('1/2/3/tableName/5/6/7');
     done();
   });
 
   it('Can fetch a lookupTable', async () => {
     let lookupTables = LookupTables.create();
     lookupTables.initialize(mockHappn);
+    await lookupTables.upsertLookupTable({
+      name: 'fetchTable',
+      paths: ['1/2/3', '4/5/6', '7/8/9']
+    });
 
-    let table = await lookupTables.fetchLookupTable('newlookuptable2');
+    let table = await lookupTables.fetchLookupTable('fetchTable');
     test.expect(table).to.eql({
-      name: 'newlookuptable2',
+      name: 'fetchTable',
       paths: ['1/2/3', '4/5/6', '7/8/9'] //Do we want leading slashes or not?
     });
     let longtable = {
@@ -169,13 +170,188 @@ describe(test.testName(__filename, 3), function() {
     test.expect(table).to.eql(longtable);
   });
 
-  it('can build permissions path from regex matches', done => {
-    let path = '/{{username}}//{{company}}/3/4/{{$1}}/5/6/{{$2}}';
-    let user = { username: 'bob', company: 'bobs buildings' };
+  it('Can upsert a group lookup permission', async () => {
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    let permission = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['on', 'get', 'set'],
+      table: 'COMPANY_ABC_LOOKUP',
+      path: '/device/{{user.custom_data.oem}}/{{user.custom_data.companies}}/{{$1}}'
+    };
+    await lookupTables.upsertLookupPermission('testGroup', permission); //Group need not exist to upsert a lookup permission. Is this an issue?
+    let stored = await mockHappn.services.data.get(
+      `/_SYSTEM/_SECURITY/_PERMISSIONS/_LOOKUP/testGroup/COMPANY_ABC_LOOKUP`
+    );
+    test.expect(stored.data.permissions[0]).to.eql(permission);
+  });
+
+  it('Can upsert multiple lookup permissions to the same group and table', async () => {
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    let permission1 = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['on'],
+      table: 'TABLE1',
+      path: '/device/{{user.custom_data.oem}}/{{user.custom_data.companies}}/{{$1}}'
+    };
+
+    let permission2 = {
+      regex: '^/_data/historianStore/device1/(.*)',
+      actions: ['get'],
+      table: 'TABLE1',
+      path: '/device/blah/blah/{{$1}}'
+    };
+    await lookupTables.upsertLookupPermission('testGroup2', permission1);
+    await lookupTables.upsertLookupPermission('testGroup2', permission2);
+
+    let stored = await mockHappn.services.data.get(
+      `/_SYSTEM/_SECURITY/_PERMISSIONS/_LOOKUP/testGroup2/TABLE1`
+    );
+    test.expect(stored.data.permissions[0]).to.eql(permission1);
+    test.expect(stored.data.permissions[1]).to.eql(permission2);
+  });
+
+  it('If we upsert the same permission multiple times we dont store it again (avoid data bloat)', async () => {
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    let permission1 = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['on'],
+      table: 'TABLE2',
+      path: '/device/{{user.custom_data.oem}}/{{user.custom_data.companies}}/{{$1}}'
+    };
+
+    let permission2 = JSON.parse(JSON.stringify(permission1));
+    await lookupTables.upsertLookupPermission('testGroup3', permission1);
+    await lookupTables.upsertLookupPermission('testGroup3', permission2);
+
+    let stored = await mockHappn.services.data.get(
+      `/_SYSTEM/_SECURITY/_PERMISSIONS/_LOOKUP/testGroup3/TABLE2`
+    );
+    test.expect(stored.data.permissions.length).to.be(1);
+    test.expect(stored.data.permissions[0]).to.eql(permission2);
+  });
+
+  it('Can fetch all of a groups lookup permissions', async () => {
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    let permission1 = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['on'],
+      table: 'testTable1',
+      path: '/device/{{user.custom_data.oem}}/{{user.custom_data.companies}}/{{$1}}'
+    };
+
+    let permission2 = {
+      regex: '^/_data/historianStore/device1/(.*)',
+      actions: ['get'],
+      table: 'testTable1',
+      path: '/device/blah/blah/{{$1}}'
+    };
+
+    let permission3 = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['get', 'on'],
+      table: 'testTable2',
+      path: '/device/another/{user.company}/{{$1}}'
+    };
+    await lookupTables.upsertLookupPermission('testGroup4', permission1);
+    await lookupTables.upsertLookupPermission('testGroup4', permission2);
+    await lookupTables.upsertLookupPermission('testGroup4', permission3);
+
+    let stored = await lookupTables.__fetchGroupLookupPermissions('testGroup4');
+    test.expect(stored).to.eql([permission1, permission2, permission3]);
+  });
+
+  it('can build permissions path from regex matches - tests __buildPermissionsPaths', done => {
+    let path = '/{{username}}/{{company}}/3/4/{{$1}}/5/6/{{$2}}';
+    let user = { username: 'bob', company: 'bobs_buildings' };
     let matches = ['$EntireMatch..', 'component1', 'method3'];
     let lookupTables = LookupTables.create();
     let pathsArray = lookupTables.__buildPermissionPaths(user, path, matches);
-    test.expect(pathsArray).to.eql(['/bob//bobs buildings/3/4/component1/5/6/method3']);
+    test.expect(pathsArray).to.eql(['/bob/bobs_buildings/3/4/component1/5/6/method3']);
     done();
+  });
+
+  it('tests that __testLookupPermission will return early if action not included in permission', async () => {
+    let path = { match: sinon.stub() };
+    let permission = { actions: ['on'] };
+    let lookupTables = LookupTables.create();
+    test
+      .expect(await lookupTables.__testLookupPermission({}, permission, path, 'get'))
+      .to.be(false);
+    test.expect(path.match.notCalled).to.be(true);
+  });
+
+  it('tests that __testLookupPermission will return early if regEx doesnt match', async () => {
+    let path = '1/2/3/4';
+    let permission = { actions: ['on'], regex: '^/_data/historianStore/device1/(.*)' };
+    let lookupTables = LookupTables.create();
+    lookupTables.__buildPermissionPaths = sinon.stub();
+    test.expect(await lookupTables.__testLookupPermission({}, permission, path, 'on')).to.be(false);
+    test.expect(lookupTables.__buildPermissionPaths.notCalled).to.be(true);
+  });
+
+  it('tests that __testLookupPermission will return false if table doesnt exist', async () => {
+    let path = '1/2/3/bob';
+    let permission = { actions: ['on'], regex: '^1/(.*)/3/{{user.name}}', table: 'testTable5' };
+    let identity = { user: { name: 'bob' } };
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    test
+      .expect(await lookupTables.__testLookupPermission(identity, permission, path, 'on'))
+      .to.be(false);
+  });
+
+  it('tests that __testLookupPermission will return true if there is a path match.', async () => {
+    let path = '1/2/3/4';
+    let permission = { actions: ['on'], regex: '^1/(.*)/3/(.*)', table: 'testTable5' , path: "1/{{$1}}/3/{{user.name}}/{{$2}}"};
+    let identity = { user: { name: 'bob' } };
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    await lookupTables.upsertLookupTable({
+      name: 'testTable5',
+      paths: ['1/2/3/bob/4', '4/5/6', '7/8/9']
+    });
+    test
+      .expect(await lookupTables.__testLookupPermission(identity, permission, path, 'on'))
+      .to.be(true);
+  });
+
+
+  it('tests that testLookupTables will return true if there is a path match in any permission.', async () => {
+    // IN PROGRESS, WILL FAIL
+    let permission1 = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['on'],
+      table: 'testTable7',
+      path: '/device/{{user.name}}/{{user.company}}/{{$1}}'
+    };
+
+    let permission2 = {
+      regex: '^/_data/otherProvider/(.*)/(.*)',
+      actions: ['get'],
+      table: 'testTable7',
+      path: '/device/blah/blah/{{$1}}'
+    };
+
+    let permission3 = {
+      regex: '^/_data/historianStore/(.*)',
+      actions: ['get', 'on'],
+      table: 'testTable8',
+      path: '/device/another/{user.company}/{{$1}}'
+    };
+    let permission = { actions: ['on'], regex: '^1/(.*)/3/(.*)', table: 'testTable5' , path: "1/{{$1}}/3/{{user.name}}/{{$2}}"};
+    let identity = { user: { name: 'bob' } };
+    let lookupTables = LookupTables.create();
+    lookupTables.initialize(mockHappn);
+    await lookupTables.upsertLookupTable({
+      name: 'testTable5',
+      paths: ['1/2/3/bob/4', '4/5/6', '7/8/9']
+    });
+    test
+      .expect(await lookupTables.__testLookupPermission(identity, permission, path, 'on'))
+      .to.be(true);
   });
 });
